Adds an UnrollAndForward pass
This pass operates on a function body and, starting from the first loop,
iteratively does the following:

1. Fully unroll the loop.
2. Scan for load ops. For each load op with a statically-inferrable access index:
  1. Backtrack to the original memref alloc
  2. Find all store ops at the corresponding index (possibly transitively through renames/subviews of the underlying alloc).
  3. Find the last store that occurs and forward it to the load.
  4. If the original memref is an input memref, then forward through any renames to make the target load load directly from the argument memref (instead of any subviews, say)
3. Apply the same logic to any remaining loads not inside any for loop.

This is needed because affine-scalrep is far too slow to use on an unrolled
convolution operator. Part of the reason is that affine-scalrep solves a small
ILP for each pair of store and load ops that it's considering forwarding to
ensure that is no intervening modification of the underlying memref. In so
doing, affine-scalrep is extremely conservative about only forwarding what it
can prove is safe to forward. Since we don't have this problem (our input
models omit complicated memref interactions), we can be more aggressive with
our forwarding logic.

This pass was inspired by makslevental@'s https://github.com/llvm/circt/pull/3967

Removing `normalize-memref` from some of the test passes was because of the info in https://discourse.llvm.org/t/understanding-the-semantics-of-reinterpret-cast/71187/1; basically, normalize-memref doesn't handle offsets well, and the micro_speech program lowers to memrefs with offsets.

PiperOrigin-RevId: 539988916