fix: fix yosys-optimizer using boolean gate mapping with abc -fast mode
Fixes https://github.com/google/heir/issues/554

It appears that our custom liberty file was not providing delays for the cells, and that was causing an assertion failure in the ABC library. Since the tfhe-rs boolean gates are a subset of the standard cell library, just restrict the standard cells using `-g CELLTYPE, ...` flags.

PiperOrigin-RevId: 619626413