Run apply-folders before canonicalize and after unrolling loops
This new pass gives an empty set of patterns to the greedy pattern
rewrite engine, which ends up applying each op's folding routine,
which simplifies the IR enough to make a normal canonicalize pass
fast.

However, this reduces some of the optimality of the final IR for some
tests, via inserting additional rotations that are not necessary.  So I
added a few additional canonicalization patterns to tensor_ext that
restore the original behavior.