feat: add tosa-to-boolean-fpga-tfhe pipeline and add bool emitter support for affine, arith ops
This uses BTreeMap for the memrefs. This allowed me to create an ordered collection of its value on emitting the result. The memrefs store Ciphertext (not &Ciphertext) because some result Ciphertexts are computed and then assigned within for loops, and the result must have a lifetime longer than the for loop scope.

PiperOrigin-RevId: 622840846