Add bgv-add-client-interface pass
This PR bundles a few changes related to enabling and end-to-end lowering
through BGV to OpenFHE.

The main obstacle here is that an input function like the `simple_sum`
test added in this PR go through type changes when being compiled to
FHE. In particular, a `tensor<32xi16> -> i16` function becomes a
`rlwe_ciphertext -> rlwe_ciphertext` function without any indication
of what the original type was that generated the ciphertext.

https://github.com/google/heir/pull/614 prepared a few changes for this, and the remaining changes are:

- Adds a `bgv-add-client-interface` pass, which inserts two functions
  that encrypt and decrypt the data types input and output to each compiled
  function.
- Adds `bgv.encrypt/decrypt` ops to support the above.

The `bgv-add-client-interface` pass can likely be generalized somehow, but I
will leave that to future work.

Part of https://github.com/google/heir/issues/494

PiperOrigin-RevId: 625480692