feat: add a secret.separator op to logically separate modules into optimize-able subunits
This PR intends to split TF2FHE models into logical sub-units (e.g. matrix mul units) to further pass to Yosys optimizer.

This also fixes a bug in --secret-distribute-generic where moving an op with a region resulted in a bug because the operand mapper did not iterate over ops in the region, so there were some op operands that were undefined.

PiperOrigin-RevId: 626037475