fix: fix secret-distribute-generic when distributing over operations that contain regions
When secret-distribute-generic tries to move `affine.parallel` loops into the previous target generics, it walks over all the operands contained within the operation. Since the operation contains regions, it's possible some of the op operands are defined within the region. In this case, we don't need to add any IRMapping values for the op operand, so we can catch this case by detecting whether the op operand was defined within a region contained within opToMove.

PiperOrigin-RevId: 627734999