fix: fix local loop variable emission in verilog
I was extremely confounded yesterday with an issue in the RTLIL importer that turned out to be a red herring and was actually because I was incorrectly wiring the hardware module for generate for loops.

Before, I declared local loop vars within the generate scope but outside the for loop, which resulted in the local loop vars holding state across the loop iterations. This resulted in, for the test case shown below, the input wires' two i16 parts to be connected via the load assignment wire. We don't expect anything but the result wire to have connections, so that caused an error in hello world.

This fixes the local loop wire declarations to happen inside the for loop, and updates the test to something slightly more nontrivial (where the ints stored in the result memref are expected to be different).

PiperOrigin-RevId: 628436267