Reenable micro speech test
At this point I think it's fair to say that the old heir-tosa-to-arith
pass is not the way we want to move forward. Rather, we will
opporuntistically fuse some of the affine loops and optimize the bodies
jointly with yosys, instead of always fully unrolling.

So this changes the heir-tosa-to-arith pipeline to simply be the prefix
of the tosa-to-boolean-tfhe pipeline, without adding the secret related
passes.

Fixes https://github.com/google/heir/issues/176