bgv/openfhe: update rotation index to be an integerattr
This change updates the rotation index / offset to be an attribute rather than an SSA value, reflecting the requirement that we must have a statically known constant rotation for the RLWE rotation operation.

We can also lower uknown tensor_ext rotations like a blind rotate or PIR style operation, but that is future work if we need it.

Fixes https://github.com/google/heir/issues/741

I'll rebase this on top of https://github.com/google/heir/pull/696

PiperOrigin-RevId: 645162459