secret: add pass to absorb memref.dealloc ops into generic
When a memref is only used internally within a secret.generic, we want the memref.dealloc op to be absorbed within that generic body, to avoid that memref from being returned from the generic.

This can happen when the full body of a generic is not unrolled (which happens in the verilog path when there are parallel loops which we write with verilog generate macro statements).

In the tosa-to-boolean-tfhe pipeline, this prevents multiple memrefs from being returned from each secret.generic

PiperOrigin-RevId: 655716568