[rtlil import] Allocate result memref with shape corresponding to original result type.
In the case of yosys-optimizing a generic with a memref result type, we want to match the allocation for the result bits to match the original shape.

Before this fix, we would always allocate a flattened bit representation. For e.g. if the result shape was memref<2xi4> we would allocate memref<8xi1>. This would result in a secret.cast that was hard to handle and we couldn't fold away casting the flattened memref to the desired shape.

Instead, we allocate the memref according to the original shape. This way we can still easily fold away stores into the flattened view of the memref and casts back to the original shape.

Note that this CL is complimentary to #810 - that PR resolves the secret.cast's while this PR allows all the memref.stores generated by yosys optimizer for the result bits to be load/stored from the original desired shape. This way our final IR doesn't have any `memref.collapse_shape/expand_shape/reinterpret_cast`'s leftover.

PiperOrigin-RevId: 658128669