Add a pass to balance a tree of operations to reduce multiplicative depth or improve parallelism.
This pass takes a tree of operations, determines the operands (by making cuts on intermediate operations where necessary) to flatten it, and balances it by recursively creating new operations that combine pairs of operands.

It handles intermediate operations by checking if intermediate operations are used more than once. If so, then make a cut and balance the intermediate operations separately. Also, treat intermediate operations used as part of another computation to be balanced as an independent operand. Note that this algorithm will never increase the number of addition/multiplication operations and with intermediate operations, may mean that it does not optimally decrease the depth of operations.

The pass currently doesn't sort the operands.

PiperOrigin-RevId: 659463207