Addresses #847
This didn't reduce the build time all that much (30 minutes -> 25 minutes on my machine), but it does reduce the build time of heir-opt alone, and makes heir-opt 61 MiB when built with optimizations, vs the current 161 MiB

Nb. this is an internal patch of https://github.com/google/heir/pull/895, which was required due to the changes to the test_utilities being incompatible with copybara transforms.

PiperOrigin-RevId: 663029648