cggi: resolve cggi lut canonicalization with vectorized inputs
@ai-mannamalai notes an issue in https://github.com/google/heir/issues/519 that canonicalizing lut2 and lut3 before running straight-line-vectorizer caused a type error because cggi.lut_lincomb expects a Variadic<LWECiphertext> and cannot be vectorized to have tensor inputs.

Updating the lincombs input and output types to Variadic<LWECiphertextLike> causes a tablegen error:
Is an operation that takes a variadic number of arguments with type TypeOrContainer unsupported?
I have a use case of a very general operation that can take a variadic number of input (it's a generic LUT op, can be a lut 2, 3, etc), and I also want it to apply elementwise to tensors. I tried updating `Variadic<InputType>` to `Variadic<InputTypeOrContainer>` and got an error that
`error: Value specified for template argument 'Variadic:type' is of type TypeOrContainer; expected type Type: LWECiphertextLike`

I updated that to AnyType and added verification to ensure that this type was an LWE Ciphertext type or a container of one.

This fixes the issue, but it is certainly bizarre that the tablegen cannot case a TypeOrContainer to an LWECiphertextLike.

PiperOrigin-RevId: 667696823