[tensor_ext] Fix canonicalization pattern so that it respects 1-D tensor constraint
The rotation normalization pattern should only apply to 1-D tensors since it extracts out the first index of the tensor dimension to perform the rotation index modulus operation. This adds a CPred to enforce the 1-D constraint.

There's also the enhancement of updating it so that it performs the modulus with the correct non-unit dimension, but I don't want to clobber the code with that condition and code until we get full layout description integration.

PiperOrigin-RevId: 683316219