mlir-to-openfhe-ckks: integrate linalg FHE kernels into the e2e pipeline
Adds a Halevi Shoup square matrix vector product benchmark! Runs in about 400 ms (compared with 3 sec I think for the unrolled naive one).

Critically, note that issue #913 is plastered in a lot of places because of the lack of layout descriptions. We treat tensors with exactly one non-unit dimension as "1-D" packed in that dimension in the ciphertext. This now happens in:
* The type converter in secret-to-ckks
* The openfhe emitter when printing constants

Also, note that if we want the corresponding BGV pipeline to do this too, we need to integrate the above type converter switch in secret-to-bgv (good opportunity to share the two).

PiperOrigin-RevId: 684977133