[polynomial] fix crash lowering elementwise polynomial.add
The occurs when naively running `--lwe-to-polynomial --polynomial-to-standard` on any testcase generated from `mlir-to-bgv` or `mlir-to-ckks`

Before, this testcase would result in a segfault from casting the result type to a polynomial type instead of a tensor. Adding the error condition in ConvertAdd then resulted in a segfault from recursive application of ConvertAdd with ConvertAny - so I also added in a small check to ensure that ConvertAny only applied when types were illegal. Otherwise, ConvertAny would replace the op, then ConvertAdd would run and fail to match, then ConvertAny would match again, etc.

PiperOrigin-RevId: 693428033