Repatriate polynomial dialect
Due to some governance discussions in upstream MLIR [1], we decided it would be appropriate to move the polynomial dialect back into the HEIR tree. We also have some plans to make minor changes to the types, for instance changing the coefficient type to AnyType and removing the coefficient modulus to support things like RNS+modarith coefficient rings, and we felt it would be simpler to just make these changes in tree rather than burden the MLIR community and core maintainers with them. Cf. https://github.com/google/heir/issues/1085

[1]: https://discourse.llvm.org/t/rfc-mlir-project-charter-and-restructuring/82896

PiperOrigin-RevId: 696112603