cggi: add pass to expand lut operations into linear lwe ops and a pbs
This pass decomposes lut3, lut2, and lut_lincomb into the lwe scalar ops (mul_sclar and add) and a pbs that refreshes the ciphertext and applies the given lookup table.

This is used to translate to APIs that expose programmable_bootstrap rather than our current backend targets that use lut3s or boolean gates.

PiperOrigin-RevId: 702739340