Add extra check in GenericOp::verify
When doing context-aware dialect conversion, sometimes the GenericOp's
body can be detached from the op while the conversion is ongoing.  When
that happens, if the framework tries to verify the generic op, say,
when attempting to dump the IR, this verification hook will fail because
the region is empty and has no blocks. This extra check allows the
verification to fail gracefully.