Add setAttributeAssociatedWith, getUndistinguishedResultAttr
setAttributeAssociatedWith is the converse of
findAttributeAssociatedWith. getUndistinguishedResultAttr is a
generalized helper to allow an op to specify attributes for multiple
results via an ArrayAttr. This also makes it so that one can have both a
single-result op with a single attr and a single-result op with an
ArrayAttr of size 1, though the former is preferred.