fix: fix secret-to-cggi secret.cast conversion
The secret.cast inserted before secret-to-cggi is meant to reconcile the operands/return types of yosys optimizer, which flattens all inputs into Nxi1 memref types.

Turns out there was a bug in here where it was needlessly dealing with the original types. Now the code deals with the type converted types, which also convert i8 to 8 ciphertexts, so the secret cast lowering becomes simpler.

Before, this pass would insert a reinterpret cast and would not properly use a collapse_shape.

PiperOrigin-RevId: 748693706