dense attrs in mod_arith.constant folder
This shallowly supports dense attributes in the fold method of
mod_arith::ConstantOp::fold, since without that the MLIR infra actually
complains that fold() fails, but just on the constant op since it is
has the ConstantLike trait.

Filed a followup to properly support dense attributes throughout
mod_arith canonicalization and folding.