[secret] Update generic op printer to print arg / result attrs inline
This updates the generic printer so that the arg-specific attrs and result attrs are printed alongside the operand or result type

For example:
```
 %0 = secret.generic ins(%arg0: !secret.secret<tensor<32x32xi16>> {tensor_ext.layout = #layout1}, %arg1: !secret.secret<tensor<32x32xi16>> {tensor_ext.layout = #layout1}) {
    ^body(%input0: tensor<32x32xi16>, %input1: tensor<32x32xi16>):
      ...
      secret.yield %4 : tensor<32xi16>
    } -> (!secret.secret<tensor<32xi16>> {tensor_ext.layout = #layout})
```

Notably, this changes the syntax of the inputs. Before, we had `ins(%ssa_val0, %ssa_val1, ... : type0, type1. ...)`. Now, it is like `ins(%ssa_val0: type0 {attr0}, %ssa_val1: type1 {attr1}, ...)`

WDYT? Before I churn the tests.

PiperOrigin-RevId: 757792071