fix: memref-global-replace unnecessarily requests IR rewrite rollback
context:
There is an upcoming deprecation from MLIR upstream to not allow rollbacks from dialect conversions anymore.

With `allowPatternRollback = false`, the memref_global_input_access test fails.

This is due to the fact that we erroneously returned a match failure when the global couldn't be completely erased (i.e. when all of its accesses were replaced with the value they accessed). However, this shouldn't trigger a rollback - we still have a rewrite.

I suppose now that I understand MLIR better I would have matched on accessors rather than the global. But IIRC I think I didn't or couldn't modify outside the scope of the accessors if I matched on them. Maybe it should be implemented as a walk instead.

PiperOrigin-RevId: 758288323