cggi-decompose-operations: new pass to decompose high level CGGI ops into lwe ops (scalar ops + PBS)
This pass was motivated by the fact that CGGIToTfheRust was decomposing bitwise xor, and, or operations into scalar ops and PBS ops, and then --emit-tfhe-rust-hl would fail to emit the PBS ops.

Instead, we moved those decompositions into a common pass so users can selectively decompose operations or not. It also subsumes --expand-lut.

The pass --emit-tfhe-rust-hl should use the high level tfhe_rust.bit(and|xor|or) ops directly.

In the future, we will decompose operations further for PBS scheduling and emit the PBS directly in the HL API.

PiperOrigin-RevId: 760655077