Implement RegionBranchOpInterface for secret.generic
This commit implements RegionBranchOpInterface for secret.generic
and modifies all sparse analyses in the codebase to account for
the fact that the dataflow framework now handles the propagation
of lattices from secret.yield to the parent generic's results.

This also required me to change the logic of setToEntryState
for most analyses. It is a bit nuanced because you have to know
when setToEntryState is called (only for block arguments), and
in our case the changes in this commit use "uninitialized" for
entry states of non-secret-typed block arguments, and pessimistic
defaults for secret-typed block arguments (e.g., fresh encryption
noise for the noise analysis).

Cf. https://github.com/google/heir/pull/1843#discussion_r2109800947
for a discussion of the nuance here.

I also simply added a decent amount of cleanup to the analysis passes
and subsequent attribute materialization. Now things are more
consistent, and attributes are persisted for all secret-typed SSA
values, function arguments, and results, whereas before many analyses
just worked on the ops within a generic and expected later passes
to move/copy attributes around.