cggi: move LUT2/LUT3 to lincomb rewrites to cggi-decompose-operations
This used to exist as a canonicalization pattern on the ops, but it was really inconvenient since most backends don't use the lincomb operation, and we would have to avoid using a canonicalizer pass after CGGI conversion to avoid manifesting that operation.

Instead, we move those rewrite patterns into the cggi-decompose-operations pass and allow an option to decompose to the lincomb level, or to the full scalar & PBS level.

PiperOrigin-RevId: 765192934