Reduce degree bounds in optimize-relin ILP
Key basis degree tracks the highest exponent of the key basis, e.g., 2
in (1, s, s^2), while the dimension of this ciphertext is 3. The relin
model starts by setting the initial degree to dimension - 1 (i.e., 1 for
normal ciphertexts), and adds degree via the dynamics constraints.

However, the range of allowed degrees was the range [0,
MAX_KEY_BASIS_DEGREE], which was 3 (i.e., dimension 4). This did not
occur when mixed-degree operands are forbidden because the only way to
re-align the operands is to apply a relin, which reduces the degree.