Add layout folding analysis to layout-optimization
LayoutFoldingAnalysis can now determine if an SSA value is "free" to
change its layout. It forward propagates "freeness" in the sense that
a Value's layout is free until some op forces it to change (such as a
linalg.reduce op). This is used in layout-propagation to allow the
hoisting step to "look backward" from an operand to determine if the
conversion it would hoist to that operand will eventually (if further
hoisted) become free.

Caveats:

- Function arguments are assumed free, which will not be true once we
  have multi-function support.
- The analysis is not updatable as the IR changes, and this is somewhat
  coupled with the structure of layout-optimization in that
  layout-optimization does a reverse walk of the IR. The combination
  of doing a reverse walk and only looking backwards for foldability
  implies the lattice will not need to be updated. But as a consequence
  any other user of the analysis will need to follow the same structure.