fix: add reassociation indices for expand_shape op
I noticed a bug when compiling the Pytorch MLP that if the reassociation map for the expand shape op that is used to canonicalize the linalg.broadcast didn't include reassociation indices.

I guess it was invoking default op builders that don't require the map, and maybe other canonicalization patterns were taking precedence? Anyway, added a failing test.

PiperOrigin-RevId: 798365675