Create torch-linalg-to-ckks pipeline
Wraps linalg canonicalizations and mlir-to-rlwe(ckks) pipelines. Intended to be
specific to the subset of linalg emitted by torch exporters, but could be
generalized further when we have a second ML frontend that lowers to linalg.

Adds an option to mlir-to-rlwe to use the new layout format, and this will
be removed once we finish migrating.

Currently does _not_ work on the mnist.mlir example, but we're persisting it
so we can have a single source of truth for the pipeline we want to test
on mnist.mlir (and others) while we fix the remaining issues.

PiperOrigin-RevId: 799195297