Add a kernel implementation testing framework
Extends ArithmeticDag to support two new operations:

- Rotation
- Extract

And then adds a new KernelImplementation lib that uses the new operations
to test a basic Halevi-Shoup matvec kernel. This also required having a
more complex leaf value type. In this case the leaf node value is a struct with
a tensor shape, along with a value that can either be an SSA value or a
literal, which itself is a variant over std::vector<double> and
std::vector<std::vector<double>>.

Warning: large sections of this were generated by an LLM
PiperOrigin-RevId: 799665887