Support EvalFastRotate in OpenFHE
- Adds new type for the precomputed digit decomposition
- Adds new ops for the fast_rotation_precompute and fast_rotation
- Adds a new pass openfhe-fast-rotation-precompute
- Emits the new ops in the pke emitter

Does not yet add this new pass to the pipeline by default (it was causing some failures and I have to investigate)

Only supports the case for non-hybrid key switching.

Lattigo's API is quite different from OpenFHE's so I don't think there will be a simple way to share code here. Maybe the step where we identify which rotations to hoist can be shared, but it seems worth waiting to decide how to do that until we have some meaningful analysis there.

Fixes #1924

PiperOrigin-RevId: 801012821