migrate halevi-shoup matvec rewrite to kernel implementation using arithmetic DAG
This also tests for correctness with an e2e test. The kernel for matvec using a larger ciphertext dimension was giving incorrect results. This was fixed with a bug in the extract_slice translation!

#vc4a
AI helped write the kernel migration (but the latency was high. It took just as long as if I were to do it by hand. /shrug/)

PiperOrigin-RevId: 803499153