Add a split-preprocessing pass
This new pass splits the main entry function into two: one that is the same as
the original function signature, but lazily packs all static plaintexts at
runtime (and then calls the second function), and a second `__preprocessed`
function that requires pre-packed plaintexts to be passed in as function
arguments.

A related change will further add pre-packing helpers as part of
add-client-interface, and then a final change will update e2e tests that should
be eagerly packing plaintexts on the client side (such as learned ML weights).

However, because the original function signature is preserved, all existing e2e
tests should continue to work with this pass integrated.

PiperOrigin-RevId: 803523060