poly approximation: emit a chebyshev poly from the C-F approx to enable chebyshev basis lowerings
The evaluation of a Chebyshev polynomial with PatersonStockmeyer's algorithm gives better accuracy with their minimax style property. But our FHE lowerings didn't utilize this since it only matched on Chebyshev basis polynomials in the IR. When using the C-F approximation method, the polynomial was always translated back to a monomial basis (because that makes it easier to compose with the interval scaling function).

This CL enables the P-S eval of the chebyshev poly by keeping the result of the C-F approx in the chebyshev basis in the IR. It does this by:
* Creating a ChebyshevBasis poly class that can compose a chebyshev basis poly with a linear polynomial in the monomial basis.
* Migrating the kernel for evaluation to the new kernel templating types

#vc4a

PiperOrigin-RevId: 806402675