test: add kernel testing with layout assignment
This updates the kernel implementation tests to use the cleartext values (not the packed values) by creating their layouts and evaluating the layout assignment on the cleartext data. This way we can test more complex layouts in generality.

This also finds and fixes a bug in the 2-d convolution layouts!

PiperOrigin-RevId: 811369543