Add an OpenFHE dialect interpreter
This PR adds the ability to run OpenFHE dialect code in an interpreted
mode, rather than generating C++ code that is then compiled by clang.

This is intended to support cases in the codebase where clang compile
times are very long due to having fully unrolled FHE kernels. We have
unrolled FHE kernels because noise analysis and ciphertext management
passes do not support loops yet, and the plan to add support for loops
is based on the HALO compiler paper (see #289).

To use the interpreter, load the `openfhe_interpreter_test` bazel macro,
and use the generated mlir-opt filename from that macro in your test
file to load the openfhe dialect IR as a ModuleOp, which is passed
to an Interpreter object.

Then a call to

```
interpreter.interpret("<func_name>", <func_args>)
```

must have as its second argument a vector of `TypedCppValue` objects
whose values correspond to the same ordered arguments of the function
being interpreted. The call to `interpret` returns a vector of
results as `TypedCppValue`. These can be chained together as
in the compiled case to produce a full execution of the
encrypt/run/decrypt flow.