layout: handle non-rank reducing collapse_shape
This PR expands support for collapse_shape in layout propagation to handle non-rank reducing collapses. What this does is build a relation from the source to dest tensor types with the info from the re-association indices (each group corresponds to a row major relation to the output dimension). Then this is pre-composed with the original relation to get the result relation for the collapse_shape op result.

PiperOrigin-RevId: 820232723