layout: relax layout conditions to lower halevi-shoup kernel
Before, we restricted convert-to-ciphertext-semantics to only lowering to the halevi-shoup kernel if the layouts on the matrix were exactly the diagonalized layout for the matrix and the row major layout for the vector. however, layout optimization may tweak the layouts of the matrix and vector and still preserve functionality of the kernel (e.g. by simultaneously swapping columns of the matrix and entries of the vector). These should be enforced by the rectify layout code in layout-optimization.

So this removes the exact layout conditions on the pass and instead only enforces that the bare minimum conditions are satisfied in order to use a halevi-shoup kernel (e.g. the dimensions are power of two, the column dimension lines up with the vector size, etc).

PiperOrigin-RevId: 820366527