[docs] update docs to match current pipelines
This is a stopgap since ideally I want a declarative pipeline option (tblgen or similar) that can define the pipeline and provide docs that are auto-generated, same as passes.

Also note that the doctest mechanism is tied to tblgen generation, so I haven't added any doctests here for the time being.

PiperOrigin-RevId: 823192117