Add a callback to `IRMaterializingVisitor` for created operations.
The `IRMaterializingVisitor` takes an optional callback function that is invoked for each MLIR operation created during the DAG traversal. This is used in `ConvertToCiphertextSemantics` to apply a "materialized" attribute to the original operation after its kernel is expanded so that patterns don't re-apply.

In my case, the matvec kernel added extract_slice's and that would confuse the extract slice pattern and emit warning errors about missing layout attrs

PiperOrigin-RevId: 823557768