implement-shift-network: de-dupe ctslot mappings to targets
Refactor `tensor_ext::Mapping` to use `DenseMap`.

The `tensor_ext::Mapping` class now uses a `DenseMap<CtSlot, CtSlot>` to store mappings from target to source `CtSlot`s. This change means that we only hold on to one possible source for each target in the mapping relation. It makes a choice to store the source that is closest to the target in the mapping to hopefully choose a unique source.

See issue #2350

PiperOrigin-RevId: 824635434