fix: adjust the ciphertext index in tensor.insert_slice lowering
In convert-to-ciphertext-semantics, we lowered a tensor.insert_slice by computing the remap that would align the source slice tensor to the desired output layout space. However, if we inserted the slice at a non-zero offset, then the resulting projection into the output layout would be shifted to a non-zero offset. So re-index the slice remap permutation to zero.

Note: the projection of the slice into the output space is computed by taking the the layout relation inferred from the insert_slice and then composing that with the destination layout. (slice -> output tensor -> ciphertext-semantic output).

However, we probably could also re-index the (slice -> output tensor) to be inserted at all zeros. Only the sizes and strides matter for the layout computation. That would be equivalent to what we are doing here (slice -> output tensor at offset and then re-index to 0)

PiperOrigin-RevId: 826037693