secret: allow plaintext linalg ops to be hoisted out of secret.generic
Fixes https://github.com/google/heir/issues/2380

However, note that in the test --canonicalize has to run before and after the distribute patterns. The initial canonicalize hoists the linalg.broadcast out of the generic, but doesn't allow the region bearing for loop to be hoisted out. distribute generic passes through the loop. Then the final canonicalize is needed to hoist the following mgmt.init out of a generic.

PiperOrigin-RevId: 828969960