fix: openfhe interpreter fails to update SSA values in loops
This caused only the first iteration to populate the SSA values inside the loop. Instead, we should just use insert_or_assign to update the SSA value map, since in general we shouldn't encounter an issue with overwriting values since it's a each variable is only assigned once within the loop scope and outside the loop.

This was found when debugging a null ciphertext in the MNIST CNN example, where a ciphertext tensor was being populated in a loop but only the first element of the resulting tensor was populated.

Also:
* adds a few other ops to the interpreter, just to bundle it in to the PR
* adds liveness analysis to delete ops if they're dead after the op that was visited
PiperOrigin-RevId: 834859984