fix: fix openfhe e2e 2-d conv test
The ISL integer library uses int types in the AST codegen. When we translate the codegen back to MLIR, we use index types, which are signless integer types in MLIR. This PR updates the MLIR translation to use signed integer types to match the semantics of the ISL library.

Fixes https://github.com/google/heir/issues/2422

PiperOrigin-RevId: 835333001