performance: skip over any zero diagonal rows in a 2-D conv
a 2-D convolution is transformed into a large matrix vector product. when we diagonalize this filter matrix to evaluate this with halevi-shoup, we find many zero diagonals. this calculates the zero diagonals from the layout relation of the expanded diagonalized filter matrix and skips over the corresponding summands in the halevi-shoup implementation.

Adds
* fuzz testing, comparing both the diagonal skipping implementation and no skipping, and the naive impl
* e2e test testing that only a restricted # of diagonals are multiplied
* unit tests verifying that the calculation of the zero diagonals in right

PiperOrigin-RevId: 842382620