Add a faster heuristic for IntegerRelation::isEqual
Fixes #2489

The main idea, in Layout/Utils.h::tryProveUnequal, is to use some known
points in the domain or range of a packing relation and test them for
identical image/preimage across the two layouts. If they differ, then
the layouts must be unequal. If all tests fail to prove a difference,
then the exact `isEqual` test is run.

This heuristic makes sense for ciphertext layout relations, but not
general relations, because, for example, we know that (0, 0) is almost
always in the domain of some element of a relation with a 2D domain (all
packed 2D tensors have a first entry), and (0, 0) is almost always in
the range of every relation (all packings have a first ciphertext with a
first slot, though that slot may be unused).

In this implementation, we test the extreme points of each dimension, as
well as some interior points. For the lenet example, the runtime of
`layout-propagation` on lenet.mlir was reduced from many minutes to
0.8760 seconds in -c opt mode and 8.8 seconds in -c dbg mode.  Note, the
original implementation was slow, not because `isEqual` is slow for one
particular call, but that `isEqual` is called many times and is mildly
slow for all of them, so the speedup is realized by making most, but not
all, tests take the fast heuristic path.